+++ modules/core/src/main/scala/ghyll/derivation/auto.scala
@@ -1,0 +1,7 @@
+package ghyll.derivation
+
+import ghyll.Decoder
+
+object auto {
+  implicit def autoDerive[A](implicit d: DerivedDecoder[A]): Decoder[A] = ghyll.deriveDecoder[A]
+}